requirements.txt
cryptography==43.0.1
pycryptodome==3.22.0
